Updated built contracts, coverage and tests

+++ contracts/mocks/tokens/MockLEND.sol
@@ -7,7 +7,7 @@ import "./MintableERC20.sol";
 
 contract MockLEND is MintableERC20 {
 
-    uint public decimals = 18;
+    uint256 public decimals = 18;
     string public symbol = "LEND";
     string public name = "LEND";
 }
