+++ contracts/mocks/tokens/MockMKR.sol
@@ -6,7 +6,7 @@ import "./MintableERC20.sol";
 
 contract MockMKR is MintableERC20 {
 
-    uint public decimals = 18;
+    uint256 public decimals = 18;
     string public symbol = "MKR";
     string public name = "Maker";
 }
