+++ contracts/mocks/tokens/MockREP.sol
@@ -6,7 +6,7 @@ import "./MintableERC20.sol";
 
 contract MockREP is MintableERC20 {
 
-    uint public decimals = 18;
+    uint256 public decimals = 18;
     string public symbol = "REP";
     string public name = "Augur";
 }
