+++ contracts/mocks/tokens/MockSUSD.sol
@@ -6,7 +6,7 @@ import "./MintableERC20.sol";
 
 contract MockSUSD is MintableERC20 {
 
-    uint public decimals = 6;
+    uint256 public decimals = 6;
     string public symbol = "SUSD";
     string public name = "Synthetix USD";
 }
