+++ contracts/mocks/tokens/MockUSDC.sol
@@ -6,7 +6,7 @@ import "./MintableERC20.sol";
 
 contract MockUSDC is MintableERC20 {
 
-    uint public decimals = 6;
+    uint256 public decimals = 6;
     string public symbol = "USDC";
     string public name = "USD Coin";
 }
