+++ contracts/mocks/tokens/MockUSDT.sol
@@ -6,7 +6,7 @@ import "./MintableERC20.sol";
 
 contract MockUSDT is MintableERC20 {
 
-    uint public decimals = 6;
+    uint256 public decimals = 6;
     string public symbol = "USDT";
     string public name = "USDT Coin";
 }
